splitted ast modules
