+++ src/Octizys/Ast/Type.hs
@@ -2,346 +2,30 @@
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# OPTIONS_GHC -Wno-unrecognised-pragmas #-}
 
-module Octizys.Ast.Type where
+module Octizys.Ast.Type (Type (TMono, TPoly), instanceType, hasTypeVar) where
 
 import Control.Arrow ((<<<))
-import Data.Foldable (foldl')
-import Data.List.NonEmpty (NonEmpty ((:|)), cons)
-import qualified Data.List.NonEmpty as NonEmpty
-import Data.Set (singleton)
-import qualified Data.Set as Set
+import Data.Aeson (ToJSON)
+import GHC.Generics (Generic, Generically (..))
+
+import Effectful (Eff, (:>))
+
+import Octizys.Ast.Type.Basics
+  ( InferenceVariable
+  , NormalizeType (normalize)
+  , TypeEq (typeEq)
+  , TypeValue
+  , TypeVariable
+  )
+import Octizys.Ast.Type.MonoType (MonoType)
+import qualified Octizys.Ast.Type.MonoType as Mono
+import Octizys.Ast.Type.Scheme (Scheme, instanceScheme)
+import qualified Octizys.Ast.Type.Scheme as Scheme
 import Octizys.Classes.FreeVariables (FreeVariables (freeVariables))
 import Octizys.Classes.From (From (from))
-import Octizys.FrontEnd.Cst.Type (TypeVariableId)
-
-import Data.Aeson (ToJSON, ToJSONKey)
-import Data.Map (Map)
-import qualified Data.Map as Map
-import qualified Data.Maybe
-import Data.Text (Text)
-import Effectful (Eff, (:>))
-import GHC.Generics (Generic, Generically (..))
-import Octizys.Common.Id (GenerateFromInt (generateFromInt))
-import Octizys.Effects.IdGenerator.Effect (IdGenerator, generateId)
+import Octizys.Effects.IdGenerator.Effect (IdGenerator)
 import Octizys.Format.Class (Formattable (format))
-import qualified Octizys.Format.Config as Format
-import qualified Octizys.Format.Utils as Format
-import Prettyprinter (Doc, Pretty, pretty)
-import qualified Prettyprinter as Pretty
-
-
-class TypeEq a where
-  typeEq :: a -> a -> Bool
-
-
-class NormalizeType a where
-  normalize :: a -> a
-
-
-data TypeValue = BoolType | IntType
-  deriving (Show, Eq, Ord, Generic)
-  deriving (ToJSON) via Generically TypeValue
-
-
-instance FreeVariables TypeVariableId TypeValue where
-  freeVariables _ = mempty
-
-
-instance Pretty TypeValue where
-  pretty BoolType = pretty @Text "Bool"
-  pretty IntType = pretty @Text "Int"
-
-
-instance Formattable TypeValue where
-  format = formatValue
-
-
-instance TypeEq TypeValue where
-  typeEq = (==)
-
-
-data InferenceVariable
-  = ErrorVariable Text
-  | RealTypeVariable TypeVariableId
-  deriving (Show, Eq, Ord, Generic)
-  deriving (ToJSON) via Generically InferenceVariable
-
-
-instance ToJSONKey InferenceVariable
-
-
-instance FreeVariables TypeVariableId InferenceVariable where
-  freeVariables (RealTypeVariable tid) = singleton tid
-  freeVariables (ErrorVariable _) = mempty
-
-
-instance Pretty InferenceVariable where
-  pretty (ErrorVariable _) = pretty @Text "ErrorVariable"
-  pretty (RealTypeVariable vid) = pretty vid
-
-
-instance Formattable InferenceVariable where
-  format _ = pretty
-
-
-instance From InferenceVariable TypeVariableId where
-  from = RealTypeVariable
-
-
-instance GenerateFromInt InferenceVariable where
-  generateFromInt sc oi i = RealTypeVariable $ generateFromInt sc oi i
-
-
-instance TypeEq InferenceVariable where
-  typeEq (RealTypeVariable x) (RealTypeVariable y) = x == y
-  typeEq _ _ = False
-
-
-newtype TypeVariable = TypeVariable' {unTypeVariable :: TypeVariableId}
-  deriving (Show, Eq, Ord, Generic)
-  deriving (ToJSON) via Generically TypeVariable
-
-
-instance Pretty TypeVariable where
-  pretty (TypeVariable' vid) = pretty vid
-
-
-instance Formattable TypeVariable where
-  format _ = pretty
-
-
-instance FreeVariables TypeVariableId TypeVariable where
-  freeVariables v = singleton v.unTypeVariable
-
-
-instance GenerateFromInt TypeVariable where
-  generateFromInt sc oi i = TypeVariable' $ generateFromInt sc oi i
-
-
-instance TypeEq TypeVariable where
-  typeEq = (==)
-
-
-inferenceVarToId :: InferenceVariable -> Maybe TypeVariableId
-inferenceVarToId (RealTypeVariable i) = pure i
-inferenceVarToId (ErrorVariable _) = Nothing
-
-
-data MonoType var
-  = VType {value :: TypeValue}
-  | Arrow
-      { start :: MonoType var
-      , remain :: NonEmpty (MonoType var)
-      }
-  | Variable var
-  deriving (Show, Eq, Ord, Generic)
-  deriving (ToJSON) via Generically (MonoType var)
-
-
-instance From (MonoType var) TypeValue where
-  from = VType
-
-
-instance From outVar inVar => From (MonoType outVar) (MonoType inVar) where
-  from VType {value} = VType {value}
-  from Arrow {start, remain} = Arrow {start = from start, remain = from <$> remain}
-  from (Variable var) = Variable (from var)
-
-
-instance Formattable var => Formattable (MonoType var) where
-  format = formatMono
-
-
-instance NormalizeType (MonoType var) where
-  normalize (Arrow start1 remain1) =
-    let
-      start2 = normalize start1
-      remain2 = normalize <$> remain1
-     in
-      flattenArrows start2 (NonEmpty.toList remain2)
-    where
-      flattenArrows :: MonoType var -> [MonoType var] -> MonoType var
-      flattenArrows initType [] = initType
-      flattenArrows initType (x : remain) =
-        case flattenArrows x remain of
-          Arrow newDom newRange ->
-            Arrow initType (NonEmpty.cons newDom newRange)
-          _ -> Arrow initType (x :| remain)
-  normalize x = x
-
-
-instance TypeEq a => TypeEq [a] where
-  typeEq [] [] = True
-  typeEq (x : xs) (y : ys) = typeEq x y && typeEq xs ys
-  typeEq _ _ = False
-
-
-instance TypeEq a => TypeEq (NonEmpty a) where
-  typeEq x y = typeEq (NonEmpty.toList x) (NonEmpty.toList y)
-
-
-instance TypeEq var => TypeEq (MonoType var) where
-  typeEq x y =
-    case (normalize x, normalize y) of
-      (VType v1, VType v2) -> typeEq v1 v2
-      (Arrow start1 remain1, Arrow start2 remain2) ->
-        typeEq start1 start2 && typeEq remain1 remain2
-      (Variable v1, Variable v2) -> typeEq v1 v2
-      (_, _) -> False
-
-
-needsParentsInArrow :: MonoType var -> Bool
-needsParentsInArrow t =
-  case t of
-    VType {} -> False
-    Arrow {} -> True
-    Variable {} -> False
-
-
-hasTypeVarMono
-  :: Eq var
-  => var
-  -> MonoType var
-  -> Bool
-hasTypeVarMono _ (VType _) = False
-hasTypeVarMono v (Arrow t1 ts) =
-  hasTypeVarMono v t1 || any (hasTypeVarMono v) ts
-hasTypeVarMono v (Variable v2) = v == v2
-
-
-arrowFromArgsAndOutput
-  :: NonEmpty (MonoType var)
-  -> MonoType var
-  -> MonoType var
-arrowFromArgsAndOutput = undefined
-
-
-replaceMonoVars
-  :: Map InferenceVariable (MonoType InferenceVariable)
-  -> MonoType InferenceVariable
-  -> MonoType InferenceVariable
-replaceMonoVars _ t@(VType _) = t
-replaceMonoVars s (Arrow t1 t2) =
-  Arrow
-    (replaceMonoVars s t1)
-    (replaceMonoVars s <$> t2)
-replaceMonoVars s t@(Variable v) =
-  Data.Maybe.fromMaybe t (Map.lookup v s)
-
-
-replaceMonoTypeId
-  :: Map TypeVariableId TypeVariableId
-  -> MonoType TypeVariable
-  -> MonoType TypeVariable
-replaceMonoTypeId _ t@(VType _) = t
-replaceMonoTypeId s (Arrow t1 t2) =
-  Arrow
-    (replaceMonoTypeId s t1)
-    (replaceMonoTypeId s <$> t2)
-replaceMonoTypeId s t@(Variable v) =
-  case Map.lookup (unTypeVariable v) s of
-    Just newId -> Variable (TypeVariable' newId)
-    Nothing -> t
-
-
-instance
-  (Eq var, Ord var)
-  => FreeVariables var (MonoType var)
-  where
-  freeVariables VType {} = mempty
-  freeVariables Arrow {start, remain} =
-    foldl'
-      (<>)
-      (freeVariables start)
-      (freeVariables <$> remain)
-  freeVariables (Variable v) = Set.singleton v
-
-
-data Scheme var = Scheme'
-  { arguments :: NonEmpty TypeVariableId
-  , body :: MonoType var
-  }
-  deriving (Show, Eq, Ord, Generic)
-  deriving (ToJSON) via Generically (Scheme var)
-
-
-instance Formattable var => Formattable (Scheme var) where
-  format = formatScheme
-
-
-instance
-  (Eq var, Ord var, From var TypeVariableId)
-  => FreeVariables var (Scheme var)
-  where
-  freeVariables s =
-    Set.difference
-      (freeVariables s.body)
-      (Set.fromList (NonEmpty.toList (from <$> s.arguments)))
-
-
-pairSchemeVars
-  :: NonEmpty TypeVariableId
-  -> NonEmpty TypeVariableId
-  -> Either () (Map TypeVariableId TypeVariableId)
-pairSchemeVars x y =
-  case go (NonEmpty.toList x) (NonEmpty.toList y) of
-    Right xs -> Right $ Map.fromList xs
-    Left _ -> Left ()
-  where
-    go [] [] = Right mempty
-    go (x1 : xs) (y1 : ys) =
-      case go xs ys of
-        Right out1 -> Right ((x1, y1) : out1)
-        Left _ -> Left ()
-    go _ _ = Left ()
-
-
-instance NormalizeType (Scheme var) where
-  normalize (Scheme' args1 body1) = Scheme' args1 (normalize body1)
-
-
-instance TypeEq (Scheme TypeVariable) where
-  typeEq (Scheme' args1 body1) (Scheme' args2 body2) =
-    case pairSchemeVars args1 args2 of
-      Right sub -> typeEq (replaceMonoTypeId sub body1) body2
-      Left _ -> False
-
-
-{- | Closes the variables of a scheme by generating new
-fresh type variables for it.
--}
-instanceScheme
-  :: IdGenerator TypeVariableId :> es
-  => Scheme InferenceVariable
-  -> Eff es (MonoType InferenceVariable)
-instanceScheme s = do
-  newIdsMap <-
-    mapM
-      ( \arg -> do
-          newId <- generateId Nothing
-          pure (from arg, Variable $ RealTypeVariable newId)
-      )
-      (NonEmpty.toList s.arguments)
-  let
-    context = Map.fromList newIdsMap
-  pure $ replaceMonoVars context s.body
-
-
-hasTypeVarScheme
-  :: InferenceVariable
-  -> Scheme InferenceVariable
-  -> Bool
-hasTypeVarScheme (ErrorVariable _) _ = False
-hasTypeVarScheme v@(RealTypeVariable vid) (Scheme' args body) =
-  notElem vid args && hasTypeVarMono v body
-
-
-instance From outVar inVar => From (Scheme outVar) (Scheme inVar) where
-  from Scheme' {arguments, body} =
-    Scheme'
-      { arguments = arguments
-      , body = from body
-      }
+import Octizys.FrontEnd.Cst.Type (TypeVariableId)
 
 
 data Type var
@@ -407,69 +91,5 @@ hasTypeVar
   :: InferenceVariable
   -> Type InferenceVariable
   -> Bool
-hasTypeVar v (TMono ty) = hasTypeVarMono v ty
-hasTypeVar v (TPoly ty) = hasTypeVarScheme v ty
-
-
--- * Format
-
-
-formatValue :: Format.Configuration -> TypeValue -> Doc ann
-formatValue _ BoolType = Format.text "Bool"
-formatValue _ IntType = Format.text "Int"
-
-
-formatMono
-  :: Formattable var
-  => Format.Configuration
-  -> MonoType var
-  -> Doc ann
-formatMono configuration VType {value} = formatValue configuration value
-formatMono configuration Arrow {start, remain} =
-  (Pretty.group <<< Format.nest configuration)
-    ( Pretty.line'
-        <> Pretty.concatWith
-          (\l r -> l <> Pretty.line <> pretty @Text "->" <> r)
-          ( prettyArg
-              <$> cons start remain
-          )
-    )
-  where
-    prettyArg ty =
-      if needsParentsInArrow ty
-        then Pretty.parens (formatMono configuration ty)
-        else formatMono configuration ty
-formatMono configuration (Variable v) = format configuration v
-
-
-formatScheme
-  :: Formattable var
-  => Format.Configuration
-  -> Scheme var
-  -> Doc ann
-formatScheme configuration (Scheme' {arguments, body}) =
-  pretty @Text "forall"
-    <> Format.nest
-      configuration
-      ( Pretty.line
-          <> Pretty.fillSep
-            ( Pretty.punctuate
-                (pretty ',')
-                (pretty <$> NonEmpty.toList arguments)
-            )
-      )
-    <> Pretty.line
-    <> pretty '.'
-    <> Format.nest
-      configuration
-      ( formatMono configuration body
-      )
-
-
-formatType
-  :: Formattable var
-  => Format.Configuration
-  -> Type var
-  -> Doc ann
-formatType configuration (TMono t) = formatMono configuration t
-formatType configuration (TPoly t) = formatScheme configuration t
+hasTypeVar v (TMono ty) = Mono.hasTypeVar v ty
+hasTypeVar v (TPoly ty) = Scheme.hasTypeVar v ty
